Add image optimisation
